feat/유저,포스트 관계설정

+++ server/src/post/post.service.ts
@@ -3,6 +3,7 @@ import { CreatePostDto } from './dto/create-post.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Post } from './post.entity';
+import { User } from 'src/auth/user.entity';
 
 /*
     service란
@@ -17,10 +18,11 @@ export class PostService {
     private postRepository: Repository<Post>
   ) {}
 
-  async getAllMarkers() {
+  async getAllMarkers(user: User) {
     try {
       return this.postRepository
         .createQueryBuilder('post')
+        .where('post.userId = :userId', { userId: user.id })
         .select(['post.latitude', 'post.longitude', 'post.type'])
         .getMany();
     } catch (error) {
@@ -29,22 +31,24 @@ export class PostService {
     }
   }
 
-  async getPosts(page: number) {
+  async getPosts(page: number, user: User) {
     const perPage = 10;
     const offset = (page - 1) * perPage;
     return this.postRepository
       .createQueryBuilder('post')
+      .where('post.userId = :userId', { userId: user.id })
       .orderBy('post.date', 'DESC')
       .take(perPage)
       .skip(offset)
       .getMany();
   }
 
-  async getPostById(id: number) {
+  async getPostById(id: number, user: User) {
     try {
       const foundPost = await this.postRepository
         .createQueryBuilder('post')
-        .where('post.id = :id', { id })
+        .where('post.userId = :userId', { userId: user.id })
+        .andWhere('post.id = :id', { id })
         .getOne();
       if (!foundPost) {
         throw new NotFoundException('게시물을 찾을 수 없습니다.');
@@ -56,7 +60,7 @@ export class PostService {
     }
   }
 
-  async createPost(createPostDto: CreatePostDto) {
+  async createPost(createPostDto: CreatePostDto, user: User) {
     const { latitude, longitude, type, address, title, description, date, score, time, imageUris } =
       createPostDto;
 
@@ -70,6 +74,7 @@ export class PostService {
       date,
       score,
       time,
+      user,
     });
 
     try {
@@ -79,16 +84,19 @@ export class PostService {
       throw new InternalServerErrorException('게시물 저장에 실패했습니다.');
     }
 
-    return post;
+    const { user: _, ...postWithoutUser } = post; // user 정보 제외
+
+    return postWithoutUser;
   }
 
-  async deletePost(id: number) {
+  async deletePost(id: number, user: User) {
     try {
       const result = await this.postRepository
         .createQueryBuilder('post')
         .delete()
         .from(Post)
-        .where('id = :id', { id })
+        .where('userId = :userId', { userId: user.id })
+        .andWhere('id = :id', { id })
         .execute();
       if (result.affected === 0) {
         throw new NotFoundException('게시물을 찾을 수 없습니다.');
@@ -102,9 +110,10 @@ export class PostService {
 
   async updatePost(
     id: number,
-    updatePostDto: Omit<CreatePostDto, 'latitude' | 'longitude' | 'address' | 'time'>
+    updatePostDto: Omit<CreatePostDto, 'latitude' | 'longitude' | 'address' | 'time'>,
+    user: User
   ) {
-    const post = await this.getPostById(id);
+    const post = await this.getPostById(id, user);
     const { title, description, score, imageUris } = updatePostDto;
     post.title = title;
     post.description = description;
